+++ src/main/ipcHandlers/settingsIpcHandlers.ts
@@ -1,0 +1,1 @@
+// ipcHandlers for setting operations
