+++ tsconfig.web.json
@@ -5,7 +5,7 @@
     "src/renderer/src/**/*",
     "src/renderer/src/**/*.tsx",
     "src/preload/*.d.ts"
-  ],
+, "src/types"  ],
   "compilerOptions": {
     "composite": true,
     "jsx": "react-jsx",
